Support Warcraft III 3.0.0 (Forsaken Kingdom) MDX 1800 and the _de.w3mod CASC layer
Format changes derived from all 14,956 stock models on the 3.0.0 CDN:
- GEOS SKIN elements are 16-bit (bone ids + weights); skin indices are int[] through the model classes
- LITE gains ShadowCasting, ShadowCastingStart/End, QuadraticFalloff, LinearFalloff, Damping
- CAMS stores a header byte in the top of the inclusive size and three depth-of-field tracks
  (IDUF focus distance, ELAF focal length, PTSF f-stop); unknown trailing camera data is skipped
- new zero-filled DILG ("Glider") chunk is preserved verbatim
- layer shading bits 0x200/0x400 (BackFacesForShadows/AmbientOcclusion, best guess) and unknown
  shader type ids (2, 24) are preserved
- versions >= 1600 use the new layouts (two stock day/night models are 1600/1700)
- MDL text read/write for the new light and camera fields

Data sources: detect 3.0 installs, add a Definitive Edition (_de.w3mod) graphics-mode preset,
count _de.w3mod as HD, drop the removed 2.0 hd2 addon preset. Menu: assign FormatVersion 1800.

Also fixes two pre-existing loader bugs surfaced by the 3.0 data: BPOS was dropped for models
without FaceFX, and a bind pose shorter than nodes+cameras crashed updateIdObjectReferences.

+++ craft3data/src/com/hiveworkshop/wc3/mdl/GeosetVertex.java
@@ -22,7 +22,7 @@ public class GeosetVertex extends Vertex {
 	List<TVertex> tverts = new ArrayList<>();
 	private List<GeosetVertexBoneLink> links = new ArrayList<>();
 	List<Triangle> triangles = new ArrayList<>();
-	private byte[] skinBoneIndexes;
+	private int[] skinBoneIndexes;
 	private float[] tangent;
 
 	Geoset geoset;
@@ -41,7 +41,7 @@ public class GeosetVertex extends Vertex {
 	}
 
 	public void initV900Skin() {
-		skinBoneIndexes = new byte[4];
+		skinBoneIndexes = new int[4];
 		while (links.size() > 4) {
 			links.remove(links.size() - 1);
 		}
@@ -86,7 +86,7 @@ public class GeosetVertex extends Vertex {
 			this.links.add(new GeosetVertexBoneLink(link.weight, link.bone));
 		}
 		if (old.skinBoneIndexes != null) {
-			this.skinBoneIndexes = new byte[old.skinBoneIndexes.length];
+			this.skinBoneIndexes = new int[old.skinBoneIndexes.length];
 			System.arraycopy(old.skinBoneIndexes, 0, this.skinBoneIndexes, 0, this.skinBoneIndexes.length);
 		}
 		this.tverts = new ArrayList<>();
@@ -218,7 +218,7 @@ public class GeosetVertex extends Vertex {
 		return tangent;
 	}
 
-	public byte[] getSkinBoneIndexes() {
+	public int[] getSkinBoneIndexes() {
 		return skinBoneIndexes;
 	}
